+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/bitset/TestDataFormatterGenericBitset.py
@@ -73,6 +73,11 @@ class GenericBitsetDataFormatterTestCase(TestBase):
         self.build(dictionary={"USE_LIBCPP": 1})
         self.do_test_value()
 
+    @add_test_categories(["msvcstl"])
+    def test_value_msvcstl(self):
+        self.build()
+        self.do_test_value()
+
     def do_test_ptr_and_ref(self):
         """Test that ref and ptr to std::bitset is displayed correctly"""
         (_, process, _, bkpt) = lldbutil.run_to_source_breakpoint(
@@ -100,4 +105,9 @@ class GenericBitsetDataFormatterTestCase(TestBase):
     @add_test_categories(["libc++"])
     def test_ptr_and_ref_libcpp(self):
         self.build(dictionary={"USE_LIBCPP": 1})
+        self.do_test_ptr_and_ref()
+
+    @add_test_categories(["msvcstl"])
+    def test_ptr_and_ref_msvcstl(self):
+        self.build()
         self.do_test_ptr_and_ref()