+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/chrono/TestDataFormatterStdChrono.py
@@ -432,6 +432,23 @@ class StdChronoDataFormatterTestCase(TestBase):
             ],
         )
 
+    @add_test_categories(["msvcstl"])
+    def test_msvcstl_durations(self):
+        self.build()
+        lldbutil.run_to_source_breakpoint(
+            self, "break here", lldb.SBFileSpec("main.cpp", False)
+        )
+        self.expect("frame variable ns", substrs=["ns = 1 ns"])
+        self.expect("frame variable us", substrs=["us = 12 µs"])
+        self.expect("frame variable ms", substrs=["ms = 123 ms"])
+        self.expect("frame variable s", substrs=["s = 1234 s"])
+        self.expect("frame variable min", substrs=["min = 12345 min"])
+        self.expect("frame variable h", substrs=["h = 123456 h"])
+        self.expect("frame variable d", substrs=["d = 654321 days"])
+        self.expect("frame variable w", substrs=["w = 54321 weeks"])
+        self.expect("frame variable m", substrs=["m = 4321 months"])
+        self.expect("frame variable y", substrs=["y = 321 years"])
+
     @skipIf(compiler="clang", compiler_version=["<", "17.0"])
     @add_test_categories(["libc++"])
     def test_libcxx(self):